8.13

+++ src/http/http.cpp
@@ -334,7 +334,7 @@ int keepAlive(Event *ev)
     return 0;
 }
 
-int  processRequestLine(Event *ev)
+int processRequestLine(Event *ev)
 {
     // LOG_INFO << "process request line";
     Connection *c = ev->c;
@@ -797,7 +797,7 @@ int runPhases(Event *ev)
 
     // OK: keep running phases
     // ERROR/DONE: quit phase running
-    while (r->at_phase < phases.size() && phases[r->at_phase].checker)
+    while ((size_t)r->at_phase < phases.size() && phases[r->at_phase].checker)
     {
         ret = phases[r->at_phase].checker(r, &phases[r->at_phase]);
         if (ret == OK)
